annotated_arg headers code refactor
