+++ sycl/include/sycl/ext/oneapi/annotated_arg/annotated_arg.hpp
@@ -30,18 +30,6 @@ struct HasSubscriptOperator
     : std::bool_constant<
           !std::is_void_v<decltype(std::declval<T>().operator[](0))>> {};
 
-// Deduce a `properties<>` type from given variadic properties
-template <typename... Args> struct DeducedProperties {
-  using type = decltype(properties{std::declval<Args>()...});
-};
-
-// Partial specialization for deducing a `properties<>` type by forwarding the
-// given `properties<>` type
-template <typename... Args>
-struct DeducedProperties<detail::properties_t<Args...>> {
-  using type = detail::properties_t<Args...>;
-};
-
 } // namespace detail
 
 // Deduction guide
